Move conversion logic to its own file
